requirements.txt
# SlideToObsidian - PPTX to AI-optimized Markdown + PDF for Obsidian
# Python 3.11+

markitdown>=0.0.1
python-pptx>=0.6.23
pymupdf>=1.24.0
tqdm>=4.66.0
